fix: make tests compatible with Node.js -pre releases
I don't know why, but `execa.node()` with the `input` option works with
Node.js -pre releases where the current configuration fails. This has
the benefit of making sure that the test is run with the version of
Node.js currently being used to execute the test suite, rather than
whatever is first in the user's path.

Before, with Node.js 13.0.0-pre:

```
$ ~/io.js/node node_modules/.bin/ava test/test.js

  1 test failed

  binary stanity

  /Users/trott/temp/csv-parser/test/test.js:247

   246:
   247:   t.snapshot(stdout)
   248: })

  Did not match snapshot

  Difference:

  - ''
  + '{"a":"1"}'
$
```